Do not generate undefined edge cases for now, plus some assumptions
Undefined edge cases like:
- Creating gaps of execution given a combination of arrival times and burst times
Ex: pid A with AT=2 and BT=2 followed by pid B with AT=5 and BT=2. From pc 4-5, nothing executes.

Current questions & assumptions:
Q1: Would having processes arrive at the same time a process finishes (regardless of whether or not the queue is empty) a problem?
A1: For now, the code is set up to try and minimize this.
Q2: Would not testing a zero offset be alright?
A2: The generator is set up to always generate non-zero offsets as the current assumption is that if the code handles an offset, it should be able to handle a 0 offset as well.