Make solving time info not reference "solve"
This is because it times everything from the click of the button up until its conclusion, so everything from accessing and getting the inputs, up to formatting and displaying the result on the DOM, is counted.

For now, it is fine to measure not just the solving part. I may decide change this in the future.