Make testcase converter display algo and quantum (if RR)
The point is so that the information about the algo and quantum to input to the website calculators is conveniently found (i.e., no need to switch to the original testcase file or check its filename). This is more useful for the non-generated testcases as they have different ATs and BTs. This in no way makes the converter become algo-specific and behave differently.